refactor(fsd): introduce views layer and thin app routes
Phase C-6. Pull every page composition out of src/app/(domain)/.../page.tsx
into src/views/<route>/<View>.tsx with a co-located index.ts barrel.
The page.tsx files are now one-line re-exports of their view's default,
so src/app/ is a thin routing adapter as required by the FSD app layer.

Views created (with index.ts barrels):
- home, signin, signup, account
- post-list, post-detail, post-create, post-update

Each view consumes features and entities through their public-API
barrels (e.g. import { SignInForm, useSignInForm } from '@features/auth-signin').

Also trim the public-API barrels to be RSC-safe:
- entities/auth/index.ts: drop fetchUser (server-only via next/headers).
- features/auth-signin/index.ts: drop fetchSignIn (server-only).
- features/auth-signup/index.ts: drop fetchSignUp (server-only).
- entities/post/index.ts: drop Editor and ContentsEditor (TUI editor
  references DOM globals at module load and breaks SSR pre-render).
Server-only and DOM-only modules are still importable via deep paths
from server actions or 'use client' files that need them.

Verified: tsc 0 errors, lint 0 errors / 233 warns (down from 287, mostly
import/order on legacy files), 7 test suites / 44 tests pass, next build
emits all 8 routes including pre-rendered /post.
