+++ src/app/(domain)/post/[id]/update/page.tsx
@@ -1,30 +1,1 @@
-'use client';
-
-import UpdatePostForm from '@features/post-update/ui/UpdatePostForm';
-import { useUpdatePostManager } from '@features/post-update/model/usePostUpdateManager';
-import Spinner from '@shared/ui/spinner';
-import LayerSpinner from '@shared/ui/spinner/LayerSpinner';
-import React from 'react';
-import { FormProvider } from 'react-hook-form';
-
-export default function UpdatePost() {
-  const {
-    methods,
-    handleUpdatePost,
-    isPending,
-    isLoading,
-    isFetching,
-    isError,
-  } = useUpdatePostManager();
-
-  if (isFetching || isLoading || isError) return <Spinner />;
-
-  return (
-    <FormProvider {...methods}>
-      <section className="mt-5 flex min-h-full flex-col justify-center border-2 border-teal-600 px-6 py-12 lg:px-8">
-        <UpdatePostForm handleUpdatePost={handleUpdatePost} />
-      </section>
-      {isPending && <LayerSpinner />}
-    </FormProvider>
-  );
-}
+export { PostUpdateView as default } from '@views/post-update';
