+++ src/features/auth-signin/index.ts
@@ -1,4 +1,3 @@
 export { default as SignInForm } from './ui/SignInForm';
 export { useSignInForm } from './model/useSignInForm';
-export { fetchSignIn } from './api/signinService';
 export type { SignInReq, SignInRes } from './api/types';
