+++ src/features/auth-signup/index.ts
@@ -1,5 +1,4 @@
 export { default as SignUpForm } from './ui/SignUpForm';
 export { default as useSignup } from './model/useSignup';
 export { useSignupForm } from './model/useSignupForm';
-export { fetchSignUp } from './api/signupServices';
 export type { SignUpReq, SignUpRes } from './api/types';
