+++ src/views/account/index.ts
@@ -1,0 +1,1 @@
+export { default as AccountView } from './AccountView';
