+++ src/views/post-create/index.ts
@@ -1,0 +1,1 @@
+export { default as PostCreateView } from './PostCreateView';
