+++ src/views/post-detail/index.ts
@@ -1,0 +1,1 @@
+export { default as PostDetailView } from './PostDetailView';
