+++ src/views/post-list/index.ts
@@ -1,0 +1,1 @@
+export { default as PostListView } from './PostListView';
