+++ src/views/post-update/index.ts
@@ -1,0 +1,1 @@
+export { default as PostUpdateView } from './PostUpdateView';
