+++ src/views/signup/index.ts
@@ -1,0 +1,1 @@
+export { default as SignUpView } from './SignUpView';
